Add files via upload
The zip file includes all agents is too big. I've download the rest of agents which are not in the following zip direct from the workshop web site.
We still got some missing parameters on the ucs file, but once we will get the relevant data to put in there - that's a piece of cake to finish it.